% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_consensus_heatmap.R
\name{consensus_heatmap}
\alias{consensus_heatmap}
\title{Generate Consensus Heatmap}
\usage{
consensus_heatmap(consensus_mat, phe = NULL, sample_names = NULL)
}
\description{
Generate Consensus Heatmap
}
