% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation_helpers.R
\name{gene_annotation}
\alias{gene_annotation}
\title{Anntate genes using NCBI gene summary}
\usage{
gene_annotation(genes_, organism_, verbose_level_ = 1, markdown = T,
  md_filename_ = NULL, html_filename_ = NULL, use_viewer = F)
}
\description{
Anntate genes using NCBI gene summary
}
