% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_consensus_heatmap.R
\name{get_order_from_consensus}
\alias{get_order_from_consensus}
\title{Generate sample order from consensus matrix}
\usage{
get_order_from_consensus(consensus_mat)
}
\description{
Generate sample order from consensus matrix
}

