% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_consensus_heatmap.R
\name{consensus_heatmap}
\alias{consensus_heatmap}
\title{Generate Consensus Heatmap}
\usage{
consensus_heatmap(consensus_mat_, phe_ = NULL)
}
\arguments{
\item{consensus_mat_}{The consensus matrix. The row and column names of this matrix will be used in the plot.
Example: \code{NMF::consensus(...)}.}

\item{phe_}{Grouping vector.}
}
\description{
Generate Consensus Heatmap
}

