% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation_helpers.R
\name{gene_annotation}
\alias{gene_annotation}
\title{Anntate genes using NCBI gene summary}
\usage{
gene_annotation(genes_, organism_, format_ = "markdown",
  use_description_ = F)
}
\description{
Anntate genes using NCBI gene summary
}

