% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation_helpers.R
\name{gene_annotation}
\alias{gene_annotation}
\title{Anntate genes using NCBI gene summary}
\usage{
gene_annotation(genes_, organism_, md_filename_ = NULL,
  html_filename_ = NULL, verbose_level_ = 1, show_in_viewer_ = T)
}
\description{
Anntate genes using NCBI gene summary
}

